responsive layout and house styles
