+++ src/style/App.scss
@@ -1,132 +1,232 @@
 $primary-color: rgba(0, 0, 0, 0.452);
-$hover-color:  #6B63B599;
-$active-color: #6B63B5;
-$font-color1: #FFFf;
+$hover-color: #6b63b599;
+$active-color: #6b63b5;
+$font-color1: #ffff;
+
+$breakpoint-xs: 360px;
+$breakpoint-sm: 767px;
+$breakpoint-md: 1023px;
+$breakpoint-lg: 1360px;
+
+@mixin media($media) {
+  @if $media == xs {
+    @media (max-width: $breakpoint-sm) {
+      @content;
+    }
+  } @else if $media == sm {
+    @media (min-width: $breakpoint-sm) and (max-width: $breakpoint-md) {
+      @content;
+    }
+  } @else if $media == md {
+    @media (min-width: $breakpoint-md) and (max-width: $breakpoint-lg) {
+      @content;
+    }
+  } @else if $media == lg {
+    @media (min-width: $breakpoint-lg) {
+      @content;
+    }
+  }
+}
 
 @mixin button-bg($bg) {
-    background: $bg;
-    &:hover {
-      background:$hover-color;
-      transition: all 0.3s ease;
-    }
-    &:active {
-      background:$active-color;
-    } 
+  background: $bg;
+  &:hover {
+    background: $hover-color;
+    transition: all 0.3s ease;
+  }
+  &:active {
+    background: $active-color;
+  }
+}
+
+body {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: rgb(20, 20, 20);
+  background-image: url(../asset/image/Background.png);
+  background-repeat: no-repeat;
+  background-position: top;
+  background-size: contain;
+  color: $font-color1;
+
+  @include media(xs) {
+    background-image: none;
   }
 
-body{
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: rgb(20, 20, 20);
-    background-image: url(../asset/image/Background.png);
-    background-repeat: no-repeat;
-    background-position:top;
-    background-size:contain;
-    color: $font-color1;
-
-    .container-btn-menu{
-        float: right;
-        margin: {
-            right: 10rem;
-            bottom: 5rem;
-        }
-        
-        .favorite-btn, .add-btn {
-            background-color:$active-color;
-            border-color: $active-color;
-            color:$font-color1;
-            width: 13rem;
-            height: 3rem;
-            font-size: 1.2rem;
-            letter-spacing: 0.2em;
-            font-weight:400;
-            
-        }
-        .favorite-btn{
-            border-bottom-left-radius: 2vh;
-            i{
-                color: #FFFf;
-                align-self: center;
-                margin: 1vh;
-                font-size:1.3rem
-            }
-        }
-        .add-btn{
-            border-bottom-right-radius: 2vh;
-            i{
-                color: #FFFf;
-                align-self: center;
-                margin: 1vh;
-                font-size:1.4rem
-            }
-        }
-        .dropdown-menu{
-            width:26rem;
-            background-color:$active-color;
-            color: $font-color1;
-            font-size: 1.3rem;
-            border-bottom-left-radius: 2vh;
-            border-bottom-right-radius: 2vh;
-
-            li{
-                margin: {
-                    top:1vh;
-                    left:2.3vh;
-                    bottom: 1vh;
-                }
-            }
-            hr{
-                margin-left: 2.1vh;
-                margin-right: 2.1vh;
-            }
-        }
+  .container-btn-menu {
+    float: right;
+    margin: {
+      right: 10rem;
+      bottom: 5rem;
     }
 
-    .main-init{
-        display: flex;
-        justify-content: center;
-        flex-direction: column;
-        align-items: center;
+    @include media(xs) {
+      position: fixed;
+      bottom: 0;
+      float: none;
+      margin: 0;
+      align-self: center;
+    }
+
+    .favorite-btn,
+    .add-btn {
+      background-color: $active-color;
+      border-color: $active-color;
+      color: $font-color1;
+      width: 13rem;
+      height: 3rem;
+      font-size: 1.2rem;
+      letter-spacing: 0.2em;
+      font-weight: 400;
+
+      @include media(xs){
+        width:23vh;
+        font-size:2vh;
         align-self: center;
-        width: 100%;
-
-        .logo{
-            margin: {
-                top: 5rem;
-                bottom:3rem;
-            }
-            width: 40vh;
-            align-self:center;
-        }
-
-        .indication{
-            align-self: center;
-            font-size: 3rem;
-            margin: {
-                top:0%;
-                bottom: 8vh;
-            }
-        }
-
-        .studients, .staff{
-            color: $font-color1;
-            border-color: $active-color;
-            width: 20rem ;
-            height: 4.5rem;
-            font-size: 1.2rem;
-            letter-spacing: 0.3em;
-            @include button-bg($primary-color);
-
-            margin: {
-                top:0vh;
-                right: 6vh;
-                left: 6vh;
-                bottom: 8vh;
-            }
-
-        }
-
+      }
     }
-    
+    .favorite-btn {
+      border-bottom-left-radius: 2vh;
+      i {
+        color: #ffff;
+        align-self: center;
+        margin: 1vh;
+        font-size: 1.3rem;
+      @include media(xs) {
+        font-size: 2vh;
+      }
+    }
+      @include media(xs) {
+        border-bottom-left-radius: 0;
+        border-top-left-radius: 2vh;
+      }
+    }
+    .add-btn {
+      border-bottom-right-radius: 2vh;
+      i {
+        color: #ffff;
+        align-self: center;
+        margin: 1vh;
+        font-size: 1.4rem;
+        @include media(xs) {
+          font-size: 2vh;
+        }
+      }
+      @include media(xs) {
+        border-bottom-right-radius: 0;
+        border-top-right-radius: 2vh;
+      }
+    }
+    .dropdown-menu {
+      width: 26rem;
+      background-color: $active-color;
+      color: $font-color1;
+      font-size: 1.3rem;
+      border-bottom-left-radius: 2vh;
+      border-bottom-right-radius: 2vh;
+
+      li {
+        margin: {
+          top: 1vh;
+          left: 2.3vh;
+          bottom: 1vh;
+        }
+      }
+      hr {
+        margin-left: 2.1vh;
+        margin-right: 2.1vh;
+      }
+    }
+  }
+  
+  .init{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    width: 100%;
+  
+
+.logo {
+      margin: {
+        top: 5rem;
+        bottom: 3rem;
+      }
+      width: 40vh;
+      align-self: center;
+      &:hover {
+        width: 42vh;
+      }
+      &:focus-visible {
+        background-color: $primary-color;
+      }
+      &:active {
+        transform: scale(0.98);
+      }
+      @include media(xs) {
+        width: 25vh;
+        margin: {
+          top: 3rem;
+          bottom: 1.5rem;
+        }
+        &:hover {
+          width: 30vh;
+        }
+      }
+    }
+
+    .indication {
+      align-self: center;
+      font-size: 3rem;
+      margin: {
+        top: 0%;
+        bottom: 8vh;
+      }
+      @include media(xs) {
+        font-size: 2rem;
+        line-height: 1.2rem;
+        margin-bottom: 4vh;
+      }
+    }
+
+    .container-btn-filter {
+      display: flex;
+      flex-direction: row;
+      @include media(xs) {
+        flex-direction: row;
+        justify-content: center;
+      }
+
+      .studients,
+      .staff {
+        color: $font-color1;
+        border-color: $active-color;
+        border-width: .5vh;
+        width: 20rem;
+        height: 4.5rem;
+        font-size: 1.2rem;
+        letter-spacing: 0.3em;
+        @include button-bg($primary-color);
+
+        margin: {
+          top: 0vh;
+          right: 6vh;
+          left: 6vh;
+          bottom: 8vh;
+        }
+        @include media(xs){
+          width: 20vh;
+          height: 4rem;
+          font-size: 1rem;
+        letter-spacing: 0em;
+          margin: {
+            right: 2vh;
+            left: 1.5vh;
+            bottom: 1vh;
+          }
+        }
+      }
+    }
+  }
 }
